Run bundler specs in another scripting container
Bundler caches a lot of things in module/class instance variables, such
as the root directory and what not. While this luckily had not yet had
any impact on the implementation, it could definitely break without
warning with a new bundler release. In fact, it does not seem possible
to implement support for path sources without a way to reset bundler
before the tests. Otherwise, the bundler root (amongst other things)
will be set according to the main puck Gemfile instead of the example
app, independent of what environment/options we supply in the specs.

This setup runs the examples in a separate scripting container, where
the bundler environment should not have been set up prematurely. A
scripting container seems like a somewhat reasonable trade-off, as it is
faster than spawning a new process, as well as being reasonably easy to
configure.

Currently, the implementation marshals inputs (using inspect/eval) and
outputs (using Marshal.dump/load) to make implementation simpler. In
theory, it should be possible to pass the variables directly, but I'm
not even sure it is possible to do this correctly from with in ruby at
this point. The obvious disadvantage with this, is that passing in
complicated objects (such as stubs) would not be possible. It obviously
also isn't possible to stub anything that the implementation uses, as
it does not even have the specs loaded.