Use bundler internal gemspec loading mechanism
Instead of trying to duplicate the logic for loading a gemspec, this
uses bundlers own `__materialize__` to do it. While the current approach
seems rather straight-forward for git and rubygem sources, it turns out
path resources are a lot less straight-forward.

The main advantage is that we are no longer running the risk of doing
almost (but not quite) the same thing as bundler. As this uses the exact
same code as bundler, it should behave as bundler, no matter what kind
of Gemfile is thrown at it. Another advantage is that if bundler adds a
new kind of source, it would just work without any additional work.

The obvious downside of relying on an internal method is that it might
break without a warning with a new release. While certainly not a good
thing, I think the current dependency on the internal file system
structure of bundler is equally bad.