Use simpler public bundler API
Digging into the bundler code, it turns out there is a much nicer
interface for obtaining the materialized gem specifications. The API
also seems to have been there since v1.0.0, so seems like a safe choice.

The only drawback of this interface is that it creates a virtual bundler
dependency, which we have to manually remove. It is unclear to me why
it does this, and even more so if there is a way to distinguish the
synthetic bundler dependency from a desired dependency on bundler. In
our particular application, a dependency on bundler is never desired,
so just removing it unconditionally seems ok.