Avoid packing project dir twice in jar
In the event that the project directory is the base path for a gem,
which will happen when a gemspec is used in the gemfile, we would pack
the project directory twice. Once under app.home in the jar, and once
under gem.home/project-version.

Even worse, not only would we pack the lib and bin directory twice, we
would also pack all other files into the jar as well, as puck currently
assumes the base path does not contain anything but the unpacked gem.
This in fact means you cannot run puck with a destination path within
the current directory for such a gem, as jar doesn't allow packing a
directory into a file within that directory.

The only drawback of this, is that things will not work as expected if
the gem has non-standard load and bin paths. However, as puck has always
assumed standard paths, I see no real problem here.