Use scripting container from bundler facade
Previously, the tests use a scripting container to allow loading a from
a different context than the main project Gemfile. However, the problem
isn't confined to tests, as the dependency resolution might want to run
for a different context compared to what was loaded by bundler.

Moving the scripting-container logic out of the tests obliged a more
thorough attention to detail concerning the scripting container.
Concretely, all data going either in to or out of the scripting
container avoids eval and goes through:

   Marshal.load(String.from_java_bytes(Marshal.dump(x).to_java_bytes))

Marshal seems like a reasonable serialization format, as it is built in
and should handle most things. The purpose of going over a java byte
array is to avoid any weird interactions due to the implicit JRuby
runtime attached to each Ruby object. In practice, the byte arrays are
still wrapped with proxy objects, so one still has to be careful.

Even exceptions go through the same treatment to try to ensure that no
references exist to the scripting container and its underlying JRuby
runtime once the contained_bundler method is finished. The scripting
container is also terminated to avoid any threads or anything from
sticking around.

The code uses parse/run directly instead of run_scriptlet, as that
allows injecting a line number so that stack traces become somewhat more
usable.