Only require included gems to be installed
It turns out that `specs_for` first evaluates all gems, except those
excluded with `--without`, which means that the whole run fails unless
all gems have been installed, even if they are not included in the end
result.

In principle, it should be possible to turn off gems by setting
`Bundler.without`, or by running `bundle --without`, but those have the
side-effect of writing a .bundle/config that keeps the groups excluded.

The only remaining solution I could find was to modify the environment
variable BUNDLE_WITHOUT, which does not seem to have any side effects.