Work around an issue in Windows
`$: << File.expand_path('../../lib', __FILE__)` from an app's bin files does not work on Windows because the path ends up as "classpath:C:/META-INF/app.home/lib". It's the same on *nix but it doesn't seem to matter that the path is prefixed with just a slash.

The behaviour of File.expand_path seems to depend on not just the actual value of __FILE__, but also on how the file it's run from was loaded. That sounds very strange and I hope I'm wrong, but that's what it looks like. It does seem like when the file is loaded either as "classpath:META-INF/app.home/bin/foobar" or just "META-INF/app.home/bin/foobar" that the path returned by File.expand_path does not come out prefixed by "classpath:/", but instead a full file system + JAR path, e.g. /home/foo/bar.jar!/META-INF/app.home/lib".

I have not tested this on Windows, I'm hoping to get some help to verify that this works around the problem. The next step is figuring out how to report this as a JRuby issue.