Add merge-archives option
merge-archive is an option to merge other jars into the new jar, instead of storing it as a nested jar. One can either pass the path to each jar in an array and each path will, in the jar, remain the same. Or one could pass it as a hash where each value represent where the jar can be found and the value where it should end up within the jar.