Make load paths absolute
Paths starting with 'classpath' should be absolute, otherwise JRuby fails to read files from inside the archive. I guess for some reason it works to require source files, but for some reason not read normal files. For example aws-sdk-v1 tries to load a json-file with a path based on it's __FILE__ path and fails.

These commands demonstrate why the paths needs to be absolute. The first fails while the second works.

$ jruby -e 'puts File.read("classpath:META-INF/LICENSE.txt")'
$ jruby -e 'puts File.read("classpath:/META-INF/LICENSE.txt")'