Handle broken JRuby 9k rewind in integration spec
It turns out that IRB reads its help message from a file inside the JAR,
and does so by first reading a line, and then rewinding the file. In
JRuby 9K, the rewind consumes the rest of the input, which means an
empty help message is produced.

Technically, rewind also doesn't work correctly in JRuby-1.7, but it is
broken in a different way, and instead of discarding the output, it just
ignores the rewind, which isn't a problem for IRB, as the first line is
discarded later anyhow.

This commit adds a detector which uses a bundled Joda-time properties
file to determine how rewind after gets is handled, and marks the IRB
spec pending when rewind discards its output.

In addition, the IRB spec is complemented with a similar spec for gem
to ensure that launching the built-in Ruby commands still works.