Add two (failing) tests for a new feature
I want to make it possible to run arbitrary code using the gems in the JAR. The idea is to make something that works like the `ruby` command: `java -jar path/to.jar ruby -e 'puts "Hello world"'`, `… ruby -v` and `… ruby path/to/code.rb` should work.

In order not to create a leaky abstraction, this should ideally be the same actual code that is run as when you run just `ruby -e …`, `ruby -v`, etc. If this is not possible perhaps it would be better to skip the "ruby" part of it and make it work like `java -jar path/to.jar -e "…"`. Since no bin files should start with a dash I think it could work.