Brute force solution to stop Bundler from changing the config during dependency resolution
`Bundler.settings[x] = y` will write a new configuration. That's just how it works. There doesn't seem to be any way around it, or any way to configure Bundler without this happening. We don't want that file to be rewritten, though. This is a brute force solution that makes a backup of the file before the dependency resolution runs, and then writes the file back afterwards.